requirements.txt
Flask
Pillow
requests
Jetson.GPIO
selenium
webdriver-manager
spidev
python-dotenv
requests